backend/requirements-dev.txt
# Development Dependencies
# Install with: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest==7.3.1
pytest-flask==1.2.0

# API Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# Database Migration
flask-migrate==4.0.4
alembic==1.11.1

# Production Server
gunicorn==20.1.0

# Development Tools
flask-debugtoolbar==0.13.1
black==23.3.0
flake8==5.0.4
pylint==2.17.4

# API Testing
requests==2.31.0
httpie==3.1.0

# Environment Variables
python-dotenv==1.0.0

# CORS Support
flask-cors==4.0.0
